requirements.txt
dj-database-url==0.4.1
Django==1.10
django-extensions==1.7.2
django-filter==0.14.0
djangorestframework==3.4.6
gunicorn==19.6.0
Markdown==2.6.6
oauthlib==1.1.2
psycopg2==2.6.2
requests==2.11.1
requests-oauthlib==0.6.2
six==1.10.0
Werkzeug==0.11.10
whitenoise==3.2.1
